Refactor console code.
Create new core::console class to hold the console munging/rendering bits code
previously scattered everywhere.

May one day be useful if we want multiple consoles with different things, but
for now there is just a singleton instance providing the normal console. The
instance should be exposed to more code in the future so the Con_XXX wrapper
functions can be whittled down.